refactor(commands): extract component tests to sibling file
